--- src/probes/outcome.rs
@@ -1,33 +1,0 @@
-use crate::core::tikibase::Tikibase;
-
-/// an issue that was identified in the Tikibase
-pub trait Issue {
-    /// provides a human-readable description of the issue
-    fn describe(&self) -> String;
-
-    /// fixes this issue, returns a description of what it did
-    fn fix(&self, base: &mut Tikibase) -> String;
-
-    /// indicates whether this issues is fixable
-    fn fixable(&self) -> bool;
-}
-
-/// a sorted list of issues
-pub struct Issues(Vec<Box<dyn Issue>>);
-
-impl Issues {
-    /// appends the given issue to this issue list
-    pub fn append(&mut self, mut new_issues: Vec<Box<dyn Issue>>) {
-        self.0.append(&mut new_issues);
-    }
-
-    /// provides the issues stored in this custom class
-    pub fn issues(self) -> Vec<Box<dyn Issue>> {
-        self.0
-    }
-
-    /// provides an empty Issues instance
-    pub fn new() -> Issues {
-        Issues(vec![])
-    }
-}
